[kube-proxy] Move ipv6 related funcs to utils pkg
